+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added `cosmic` module
 - Enabled `autosquash` by default when rebasing
 - Changed to `nixos-21.05` for GNOME 40 to be able to run COSMIC
 - Removed `prezto` `utility` module as it breaks `cp` and `mv` with `sudo`